requirements-test.txt
pytest==7.4.3
pytest-flask==1.3.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1 